Move SR to Api project
