+++ src/Database/Api/SR.cs
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-namespace Monai.Deploy.InformaticsGateway.Database.EntityFramework.Configuration
+namespace Monai.Deploy.InformaticsGateway.Database.Api
 {
     public static class SR
     {
@@ -23,6 +23,7 @@ namespace Monai.Deploy.InformaticsGateway.Database.EntityFramework.Configuration
         /// Name of the key for retrieve database connection string.
         /// </summary>
         public const string DatabaseConnectionStringKey = "InformaticsGatewayDatabase";
+        public const string DatabaseNameKey = "DatabaseName";
 
     }
 }
